+++ lib/src/ast/ast.h
@@ -1,10 +1,10 @@
 #pragma once
 
+#include "Lexer/Token.h"
 #include "ast/ast_fwd.h"
 #include "ast/visitors/visitor.h"
 #include "common/value.h"
 #include "core/types.h"
-#include "lexer/Token.h"
 #include "parser/type_specifier.h"
 #include "types/type.h"
