+++ lib/test/ut/lexer/lexer_test.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "lexer/Lexer.h"
+#include "Lexer/Lexer.h"
 
 #include <array>
 #include <string_view>
